Fixes #10 1. Rename module from wp-admin to wpadmin due to errors in the build process. 2. Add license to wpadmin.py 3. Remove starttheme and startplugin from within a class to standalone functions since they were staticmethods anyways. 4. Refactor main 5. Convert README from markdown to rst. 6. create setup.py
